fix: export blob errors (#33)
